Updated channel handeling in operators
We can not rely on ThreadLocal, since we're working with tasks that might
not give us the same threads as before. This made the 'values' of thread
local to just increase over time. Instead, we let the operations get a
new connection and dispose it self. It might be a bit of a penelty performance
wise, but it is easier to follow.