Adding extra reconnect attempt when recovery if false
Even though it is recommended to use AutoRecovery and TopologicRecovery, we
want to give users who has these features activated a good experience. In the
case that there is only one host and it becomes unavailable, we try to reconnect
instantly and if we get a BrokerUnreachableException, we wait for five seconds
and try again.