(#129) Expose Mandatory Option For Publish
The mandatory option has been exposed to RawRabbit by adding a new
property to the PublishConfiguration and a new method,
WithMandatoryDelivery, on PublishConfigurationBuilder. This method also
allows the user to pass a function to the BasicReturn event handler.
Publisher has been updated to use the madatory flag in the BasicPublish
method and the BasicReturn event.
The Rider IDE user files have also been ignored.